scripts: bootimg-sota-efi: allow dynamic bootable partition
Remove the hardcoded bootable flag on the first partition and instead
search for active partitions when defining the bootable one.

This allows the ESP partition to be created as second or third partition
(for cases where the boot firmware partition needs to be defined first).

Can be reverted once updated to LmP v90.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>