conf: lmp-partner-custom: move to linux-mtk kernel
We should leverage linux-mtk recipe as it dynamically adds
quite a few .cfg files dedicated to each piece of HW.

Instead, append the kernel-meta information with standard.scc
to make sure we are compatible with LmP (zram and cgroups).

This keeps our changes to a minimum while leveraging the mtk
kernel implementation.

Also, set DISTRO_CODENAME to fix libmali build.

Signed-off-by: Michael Scott <mike@foundries.io>