linux-qcom-custom: add modsign distro features support
The qcom kernel is not based on the yocto kernel and becasue of that we can't
use the kernel features modsign.scc available on the lmp-kernel-cache that was
required by the recipes-kernel/linux/linux_ima.inc.
To fix the above we need to add the corresponding kernel config fragment
signing.cfg for the kernel module signing.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>